LayoutTests/storage/indexeddb/modern/leak-1-expected.txt

This tests that certain IDB object relationships don't cause leaks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "leak-1"
Initial upgrade needed: Old version - 0 New version - 1
PASS transactionObserver.wasCollected is true
PASS requestObserver.wasCollected is true
PASS versionChangeTransactionObserver.wasCollected is true
PASS databaseObserver.wasCollected is true
PASS openRequestObserver.wasCollected is true
PASS objectStoreObserver.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
